+++ CHANGELOG.md
@@ -185,7 +185,7 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 - Introducing extra dependencies `iwopy[scipy]`, `iwopy[pymoo]`,
 `iwopy[pygmo]`, and `iwopy[opt]` for installation of all supported optimization packages
-- Introducing the concept of optimization suites, with abstract base class `Suite`
+- Introducing the concept of optimization pipelines, with abstract base class `Pipeline`
 - Small improvements of optimizer `GG`
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/iwopy/commits/v0.5.0](https://github.com/FraunhoferIWES/iwopy/commits/v0.5.0)